Move cache globals to cache.py
